will this work

+++ WishList/Controllers/ItemController.cs
@@ -1,21 +1,31 @@
 ﻿using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public ItemController(ApplicationDbContext context)
+        public ItemController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
+            var user = _userManager.GetUserAsync(HttpContext.User);
+            var model = from item in _context.Items.ToList()
+                        where item.Id == user.Id
+                        select item;
 
             return View("Index", model);
         }
@@ -27,8 +37,9 @@ namespace WishList.Controllers
         }
 
         [HttpPost]
-        public IActionResult Create(Models.Item item)
+        public async Task<IActionResult> Create(Models.Item item)
         {
+            item.User = await _userManager.GetUserAsync(HttpContext.User);
             _context.Items.Add(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
@@ -37,9 +48,17 @@ namespace WishList.Controllers
         public IActionResult Delete(int id)
         {
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
-            _context.Items.Remove(item);
-            _context.SaveChanges();
-            return RedirectToAction("Index");
+            var user = _userManager.GetUserAsync(HttpContext.User);
+            if (user.Id == item.Id)
+            {
+                _context.Items.Remove(item);
+                _context.SaveChanges();
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return Unauthorized();
+            }
         }
     }
 }
